Relocate Abstract tests.
